Rename react_statics
