+++ lib/src/react_client/dart_interop_statics.dart
